feat: Add selection logging for ontology recommendations
Implements a mechanism to log user selection behavior for ontology terms. This data will be used to improve the relevance and accuracy of future recommendations.

The logging includes details about the selected term, ignored recommendations, and contextual information about the EML element being annotated. The data is sent to a local endpoint using `navigator.sendBeacon`.

Additionally, the README has been updated with a new banner and deployment instructions.